Update .vimrc
Ever the habitual tinkerer, I found the configuration options for
Startify (the Vim opening screen plugin). Apparently you can customize
the header message. I got a LOTR ASCII art piece and put it in.

I also deleted the line that was supposed to make Vim copy to the system
clipboard.