Improve Compatibility with VSCode and Lisp
For now I decided to wrap my init.vim file in an if statement that will
only allow it to load if it isn't being loaded by the VSCode-NeoVim
extension. I will escape commands needed inside VSCode as needed. As a
bonus I don't need to include all of those conditionals on the plugins. I also
added some options to the QuickRun Plugin to run Lisp, JavaScript, and
Rust files. Finally, the VLime server should work better for Lisp now.