Improve Zsh config
I changed my Zsh prompt theme to Agnoster. I think I'm really starting
to like that theme.

Removed the plugin to automatically activate and deactivate Python
environments. It would have been cool and useful but I couldn't get it
to work and didn't care enough to spend a large amount of time tracking
down a bug in the source code.