Remove lazygit config and add a new command
I wasn't using lazygit at all, I just liked the layout of gitui and the
fact that it was in Rust more so I deleted my lazygit config. Secondly,
I added a command to my .personal_commands file to convert a file or
directory name into snake case. It may be a complicated enough problem
to warrant a Python script.