Add zsh aliases for Pandoc
I'm using Pandoc a lot now to convert all my school documents written
with NeoVim in Markdown to pretty PDFs and HTML to turn in. It struck me
as a little silly that I was still using the full names of the commands
and instead aliased the pandoc command and some of the common tasks I do
to some short commands.