Add Barbar plugin and config
I added the Barbar plugin to make my tabs look fancy. Of course, I also
had to add the config. Also added a bunch of Barbar keymappings to
Which-Key.

The reason that this commit looks so big and has so many changed files
is that I ran stylua on every Lua file. I was tired of my auto-format
slowly changing them on by one as I edited them so I decided to end the
formatting once and for all.