Add custom nvim command to make file on save
I write a lot of my documents in Markdown, which I use Pandoc to convert
to PDFs when I need to turn in school assignments. I like to keep an eye
on the PDF document when I'm working. There were two problems with my
previous setup:

1. I had to enter a shell command in nvim whenever I wanted to Make the
   file
2. Preview (macOS's default PDF viewer) doesn't reload PDFs when they
   change

I was able to fix both of those things by first creating a custom
commmand to create an autocommand to make the file I'm working on
whenever I save it, and secondly just getting a better PDF
viewer--Skim--which autoupdates whenever the file changes. Now I can do
a sort of live preview while I'm writing in Pandoc flavored Markdown and
see the output whenever I save the file.