Search in files with fzf and new Zettelkasten plugin
I added a few new functions to the personal_commands file to
interactively search within files using ripgrep as the backend for speed
and fzf as the pretty frontend. This way I get the best of both worlds.

I also added the Telekasten plugin to my NeoVim config to help as I
transfer my Zettelkasten to Obsidian and use NeoVim as my editor.