feat: New plugin to show better code context
Added a new plugin called nvim-navic that connects to the current LSP
and gets gets the current context of the path to the line of code I'm on
like VSCode. I integrated it with Lualine to show it in my tabline. It's
better than treesitter-context, so I deleted that plugin because they
were too redundant.