fix: New duck plugin, rename aliases
Made a couple of minor changes. Installed a duck plugin which creates a
little emoji that walks around the screen. Also moved the location of my
Neovim nightly binary that I built from source so I had to change the
alias. Finally I renamed the aliases file to have an sh extension so
Neovim can figure out the filetype.