fix: LSP problems and fancy undercurls in terminal
I was having problems with the node process under Neovim using a massive
amount of CPU whenever I opened a Markdown file. I finally tracked the
problem down to the TailwindCSS language server which was enabled for
Markdown files for some reason. It has now been disabled for Markdown
and the CPU usage is back to the low number it should stay at. Also
disabled spelling autocomplete. Wasn't using it enough to justify the
distraction. Changed my TERM variable to xterm-kitty so Neovim would
know it could use the fancy undercurls in the TUI for diagnostics.