fix: Add fzf native extension to Packer and remove navic
Fixed the Telescope error that the extension fzf-native didn't actually
exist by adding fzf-native to Packer. Also removed the navic extension
that shows the code levels in the statusline. I thought it might be cool
but I wasn't using it enough for the extension to be worth it. Changed
the Telekasten home directory to my new Zettelkasten home.