refactor(nvim): remove barbecue and disable ts for long markdown files
Removed the barbecue plugin that shows me my context in the file because
it wasn't useful since my mouse is disabled and I can't click on any of
it. The Markdown Treesitter parser can take quite a while on large
Markdown files so I disabled it when the file grows too large. Made a
few other refactors like lazy loading ZenMode and Twilight when I type
those commands and changing a Which-Key mapping from a Lua function to a
string.