fix(nvim): <leader>y copies register to clipboard
Instead of trying to have a keymap to directly copy a selection to the
clipboard instead use all the native keymaps to copy something to the
default register and then use <leader>y to copy to the system clipboard.
Now I get the best of both worlds. Easily copy text to the system
clipboard out of Neovim while avoiding copying absolutely everything to
the system clipboard and polluting it.