refactor(sketchybar): move bar to top with notch
I got a new MacBook with the notch. Putting the bar at the bottom, while
I enjoy the style, doesn't make sense in terms of screen real estate
because you can't put windows in the top slice. Moved the bar back up to
the top, changed some of the styling (there's a visible background now)
removed the wifi icon, wifi popup, disk percentage used for now, and
anchored some of the items to the notch.