feat(skhd): use modes for key combos
I have a few common actions that I want do with a list of common apps
that I would like to have bound to a keyboard shortcut. Namely, focus
apps or open a new window in an app that I commonly use. Instead of
definining a long list of keyboard shortcuts, it leads to a lot less
conflict if you have one common prefix for an action, followed by an
argument for what app to perform the action on.

I'm using skhd modes for this. To open a new window, you hit alt + cmd
and `n` and then hit the key for the app you want to open a new window,
either `t` for terminal or `b` for browser. I also did the same for
focusing an app, but with the prefix alt + cmd and `f`.